+++ Converter.java
@@ -16,4 +16,12 @@ public class Converter {
         return result;
     }
 
+    private static int intExponentiation(int base, int exponent) {
+        int result = 1;
+        for (int i=0; i<exponent; i++) {
+            result *= base;
+        }
+        return result;
+    }
+
 }